+++ app/build.gradle
@@ -63,7 +63,7 @@ dependencies {
     implementation "androidx.recyclerview:recyclerview-selection:1.1.0"
 
     implementation 'androidx.recyclerview:recyclerview:1.1.1'
-
+    implementation "androidx.fragment:fragment-ktx:1.1.0"
