+++ app/src/main/java/com/example/firstapp/adapter/summorSpell/ItemsViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.firstapp.adapter.summorSpell
+
+import android.graphics.drawable.Drawable
+
+data class ItemsViewModel(val imageButton: Int, val imageText: Int, val dialogText: Int, val dialogIcon: Int, val dialogTitle: Int)
+
+
